# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Scan -\u003e Plan -\u003e Approve -\u003e Fix -\u003e Rollback REST API for golangci-lint issue triage.",
        "title": "Golangci Lint Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/fixes": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rule BR-3: the plan must be \"approved\". Rule BR-1: one scan or fix at a time per repo+branch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fixes"
                ],
                "summary": "Apply an approved fix plan",
                "parameters": [
                    {
                        "description": "Approved plan to apply",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createFixRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "plan not approved, or a scan/fix already running for this repo+branch",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/fixes/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fixes"
                ],
                "summary": "Get a fix application's result by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fix history ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.FixHistory"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
...
        "/plans": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cache hit (same issue batch already planned) returns 200 with the ready plan immediately; a cache miss enqueues generation and returns 202.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Request an AI-generated fix plan",
                "parameters": [
                    {
                        "description": "Issue IDs to plan a fix for",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cache hit -- plan already ready",
                        "schema": {
                            "$ref": "#/definitions/storage.FixPlan"
                        }
                    },
                    "202": {
                        "description": "cache miss -- plan generation enqueued",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/plans/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Get a fix plan by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.FixPlan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/plans/{id}/approve": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rule BR-3: applying a fix requires an approved plan. Rule BR-5: high-risk/breaking-change plans require explicit confirmed=true.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Approve or reject a fix plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "approve defaults to true; set false to reject",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.approvePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.FixPlan"
                        }
                    },
                    "400": {
                        "description": "confirmation required for high-risk/breaking plans",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "plan is not pending an approval decision",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.approvePlanRequest": {
            "type": "object",
            "properties": {
                "approve": {
                    "description": "Approve defaults to true (JSON omitted -\u003e zero value -\u003e approve) --\nthe endpoint's own name is \"approve\"; explicit {\"approve\": false} is\nhow a reject happens (06-api-design.md has no separate reject route).",
                    "type": "boolean"
                },
                "confirmed": {
                    "type": "boolean"
                }
            }
        },
        "api.createFixRequest": {
            "type": "object",
            "required": [
                "plan_id"
            ],
            "properties": {
                "plan_id": {
                    "type": "string"
                }
            }
        },
        "api.createPlanRequest": {
            "type": "object",
            "required": [
                "issue_ids"
            ],
            "properties": {
                "issue_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "golangci_backend_frontiir_utils.RestErr": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable business error code (e.g. \"SLA_EXPIRED\")",
                    "type": "string"
                },
                "data": {
                    "description": "Optional structured payload (e.g. conflicting ticket detail)"
                },
                "message": {
                    "description": "User-friendly error message",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        },
        "storage.FixHistory": {
            "type": "object",
            "properties": {
                "applied_by": {
                    "type": "string"
                },
                "branch_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "diff_ref": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "string"
                },
                "post_fix_scan_id": {
                    "type": "string"
                },
                "pre_fix_scan_id": {
                    "type": "string"
                },
                "result": {
                    "description": "applying|passed|failed -- \"applying\" added in M4, apply is worker-queued like scan/plan",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.FixPlan": {
            "type": "object",
            "properties": {
                "acceptance_criteria": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "after_snippet": {
                    "type": "string"
                },
                "approved_at": {
                    "type": "string"
                },
                "approved_by": {
                    "description": "ApprovedBy/ApprovedAt implement Rule BR-3's \"approver identity +\ntimestamp recorded\" requirement (10-business-rules.md) -- absent\nfrom the original M1 schema. ApprovedAt is nil until approve/reject.",
                    "type": "string"
                },
                "batch_fingerprint": {
                    "description": "BatchFingerprint/PromptVersion implement the Risk register's own\nmitigation (\"cache plan by issue+source fingerprint; store model +\nprompt version on FixPlan\") -- absent from the original M1 schema.",
                    "type": "string"
                },
                "before_snippet": {
                    "type": "string"
                },
                "breaking_change": {
                    "type": "boolean"
                },
                "code_context": {
                    "description": "CodeContext/FixStrategyCode/BeforeSnippet/AfterSnippet/SideEffects/\nImpactAnalysis/RecommendedTestCommands/AcceptanceCriteria added\n2026-08-19 to match before-fixed/*.md's section depth (§2, §4-9) --\nsee planner.PlanResult's matching doc comments for what populates each.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_behavior": {
                    "description": "required by 04-prd.md FR-4 / 05-ui-design.md's Plan Viewer -- missing from the original M1 schema",
                    "type": "string"
                },
                "current_behavior_my": {
                    "type": "string"
                },
                "files_impacted": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "fix_strategy_code": {
                    "type": "string"
                },
                "generated_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "impact_analysis": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.LintIssue"
                    }
                },
                "prompt_version": {
                    "type": "string"
                },
                "recommended_fix": {
                    "type": "string"
                },
                "recommended_fix_my": {
                    "type": "string"
                },
                "recommended_test_commands": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "risk_level": {
                    "description": "low|medium|high",
                    "type": "string"
                },
                "root_cause": {
                    "type": "string"
                },
                "root_cause_my": {
                    "description": "RootCauseMy/CurrentBehaviorMy/RecommendedFixMy are Burmese counterparts\nto the 3 fields above, for the UI's language toggle -- additive, the\nEnglish fields are unchanged and remain the fallback when these are empty.",
                    "type": "string"
                },
                "side_effects": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "description": "generating|pending|approved|rejected|applied|failed -- \"generating\"\nand \"failed\" added in M3 for the async plan-job sub-states (a client\npolling GET /plans/:id needs to tell \"still working\" from\n\"permanently failed, e.g. AI Layer unavailable\" apart); the other 4\nare 08-validation-rules.md's originally decided enum.",
                    "type": "string"
                },
                "test_plan": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.LintIssue": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "line": {
                    "type": "integer"
                },
                "linter": {
                    "type": "string"
                },
                "message": {
                    "description": "golangci-lint's issue text, minus the rule prefix -- see parser.Parse",
                    "type": "string"
                },
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.FixPlan"
                    }
                },
                "reason_my": {
                    "description": "Burmese \"why this breaks the rule\" explanation -- see parser.reasonMy",
                    "type": "string"
                },
                "rule": {
                    "type": "string"
                },
                "scan_id": {
                    "type": "string"
                },
                "severity": {
                    "description": "critical|high|medium|low|info",
                    "type": "string"
                },
                "status": {
                    "description": "open|planned|fix_applied|resolved|reopened|ignored",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}